Initial commit
After learning about the data, the Deep Learning Network was used with the Keras library in python. Keras is a high level neural networks  API  that  is  written in Python  and  developed with a focus on enabling fast experimentation”. The sequential deep learning model in Keras was used since that is best suited to classify human actions with limited knowledge.